pyproject.toml
[tool.poetry]
name = "icml2022-mips"
version = "0.1.0"
description = ""
authors = ["usaito <ys552@cornell.edu>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
obp = "^0.5.4"
scikit-learn = "1.0.2"
pandas = "1.3.5"
scipy = "1.7.3"
numpy = "^1.22.4"
matplotlib = "^3.5.2"
seaborn = "^0.11.2"
hydra-core = "1.0.7"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = 'black'
src_paths = ['real', 'synthetic']
line_length = 88
lines_after_imports = 2
force_single_line = 'True'
force_sort_within_sections = 'True'
order_by_type = 'False'
